Started work on programmable bindings.
First task: create support for programming bindings between models using
bindings with hardcoded behavior. This is likely a scaffolding process,
because the goal is about learning what bindings are and how they should
be programmed in a code-extrinsic manner.
The first bindings, of such kind of binding, are, each, supposed to bind
the background color between a source- and a target model, such that
changes in the background color of the source model is cloned in the
target model.

Signed-off-by: Jakob Brandsborg Ehmsen <jakobehmsen.github@gmail.com>