Added new kind of model: Button.
This kind may be replaced in the future by supporting simulations of
buttons in the environment by extending the range of changes sendable
from text models - or other possible future models, which could be used
for simulating buttons such as images.

It is now possible to plot buttons and create buttons from text.

Signed-off-by: Jakob Brandsborg Ehmsen <jakobehmsen.github@gmail.com>