Continued work on view/filtering.
Seems to work somewhat so far. The implemented has been simplified in
functionality to make a first revision faster: there is (temporarily) no
longer a garantee as to the last model added being the one placed
front-most.
The implementation must be tested more - and the implementation must
written better; refactored.

Signed-off-by: Jakob Brandsborg Ehmsen <jakobehmsen.github@gmail.com>