Started work on a new tool: Scale.
This tool is designed to be used for recursive proportional resizing of
models. I.e. not only the immediate model is resized, but also its
content (if any) with propertions kept - recursivly, and the content of
the content (if any) and so forth. Here, content also includes text,
meaning the font size also changes propertionally. In the future, this
could include images and other kind of content.

This should give the appearance of stretching/shrinking models and is
supposed to be used for screen economics.

Signed-off-by: Jakob Brandsborg Ehmsen <jakobehmsen.github@gmail.com>