Applied new usage of branching:
A branch is created, one or more executions can be associated to that
branch.

The executions are not necessarily performed immediately - and a branch
cannot be absorbed before it is marked as closed. After a branch has
been marked as closed, no more executions can be associated to that
closed branch.

Signed-off-by: Jakob Brandsborg Ehmsen <jakobehmsen.github@gmail.com>