Added scale command.
Added usage of scale command in scale tool.

The above was done to support relative scaling in such a way it follows
isolated model history better. More specifically, the x and y
coordinates are set relative to the current x and y, of the scaled
model, through deltas instead of absolute coordinates. In addition, a
scale transaction consists only of a scale command instead of many
effect commands.

Signed-off-by: Jakob Brandsborg Ehmsen <jakobehmsen.github@gmail.com>